+++ test/test_simple_example.cpp
@@ -1,4 +1,4 @@
-#include "concore2full/async_oper_state.h"
+#include "concore2full/spawn.h"
 #include "concore2full/global_thread_pool.h"
 #include "concore2full/profiling.h"
