+++ test/test_spawn.cpp
@@ -1,4 +1,4 @@
-#include "concore2full/async_oper_state.h"
+#include "concore2full/spawn.h"
 
 #include <catch2/catch_test_macros.hpp>
